fin de la programatio

+++ battery/Firmware/Core/Inc/utils.h
@@ -8,6 +8,6 @@
 #ifndef INC_UTILS_H_
 #define INC_UTILS_H_
 
-int read_analog_input(void);
+float read_analog_input(void);
 
 #endif /* INC_UTILS_H_ */
